♻️  API/internal package split
