+++ src/main/java/io/strimzi/kproxy/internal/InterceptorProviderFactory.java
@@ -14,12 +14,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.strimzi.kproxy.interceptor;
+package io.strimzi.kproxy.internal;
 
 import java.util.Collections;
 import java.util.List;
 
 import io.netty.channel.socket.SocketChannel;
+import io.strimzi.kproxy.interceptor.Interceptor;
 
 public class InterceptorProviderFactory {
