+++ src/main/java/io/strimzi/kproxy/internal/KafkaProxyBackendHandler.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.strimzi.kproxy;
+package io.strimzi.kproxy.internal;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
