+++ src/main/java/io/strimzi/kproxy/internal/KafkaProxyFrontendHandler.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.strimzi.kproxy;
+package io.strimzi.kproxy.internal;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -41,7 +41,6 @@ import io.strimzi.kproxy.codec.KafkaRequestEncoder;
 import io.strimzi.kproxy.codec.KafkaResponseDecoder;
 import io.strimzi.kproxy.interceptor.HandlerContext;
 import io.strimzi.kproxy.interceptor.Interceptor;
-import io.strimzi.kproxy.interceptor.InterceptorProvider;
 
 public class KafkaProxyFrontendHandler extends ChannelInboundHandlerAdapter {
